backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
langgraph
requests
numpy
soundfile
webrtcvad-wheels